@@ -5,6 +5,7 @@ import java.util.Collections;
 import java.util.List;
 
 import com.facebook.react.ReactPackage;
+import com.facebook.react.bridge.JavaScriptModule;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
@@ -13,9 +14,8 @@ import com.himelbrand.ble.peripheral.RNBLEModule;
 public class RNBLEPackage implements ReactPackage {
 
   @Override
-  public List<NativeModule> createNativeModules(
-  ReactApplicationContext reactContext) {
-    List<NativeModule> modules = new ArrayList<NativeModule>();
+  public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
+    List<NativeModule> modules = new ArrayList<>();
     modules.add(new RNBLEModule(reactContext));
     return modules;
   }
