@@ -12,25 +12,30 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
     var manager: CBPeripheralManager!
     var startPromiseResolve: RCTPromiseResolveBlock?
     var startPromiseReject: RCTPromiseRejectBlock?
-    
+
     override init() {
         super.init()
         manager = CBPeripheralManager(delegate: self, queue: nil, options: nil)
         print("BLEPeripheral initialized, advertising: \(advertising)")
     }
-    
+
     //// PUBLIC METHODS
+
+    @objc func isAdvertisementSupported(_ resolve: RCTPromiseResolveBlock, rejecter reject: RCTPromiseRejectBlock) {
+        resolve(true)
+        print("called isAdvertisementSupported")
+    }
 
     @objc func setName(_ name: String) {
         self.name = name
         print("name set to \(name)")
     }
-    
+
     @objc func isAdvertising(_ resolve: RCTPromiseResolveBlock, rejecter reject: RCTPromiseRejectBlock) {
         resolve(advertising)
         print("called isAdvertising")
     }
-    
+
     @objc(addService:primary:)
     func addService(_ uuid: String, primary: Bool) {
         let serviceUUID = CBUUID(string: uuid)
@@ -44,7 +49,7 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
             alertJS("service \(uuid) already there")
         }
     }
-    
+
     @objc(addCharacteristicToService:uuid:permissions:properties:data:)
     func addCharacteristicToService(_ serviceUUID: String, uuid: String, permissions: UInt, properties: UInt, data: String) {
         let characteristicUUID = CBUUID(string: uuid)
@@ -55,13 +60,13 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
         servicesMap[serviceUUID]?.characteristics?.append(characteristic)
         print("added characteristic to service")
     }
-    
+
     @objc func start(_ resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
         if (manager.state != .poweredOn) {
             alertJS("Bluetooth turned off")
             return;
         }
-        
+
         startPromiseResolve = resolve
         startPromiseReject = reject
 
@@ -71,7 +76,7 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
             ] as [String : Any]
         manager.startAdvertising(advertisementData)
     }
-    
+
     @objc func stop() {
         manager.stopAdvertising()
         advertising = false
@@ -98,7 +103,7 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
             alertJS("service \(serviceUUID) does not exist")
         }
     }
-    
+
     //// EVENTS
 
     // Respond to Read request
@@ -175,7 +180,7 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
         startPromiseResolve!(advertising)
         print("advertising succeeded!")
     }
-    
+
     //// HELPERS
 
     func getCharacteristic(_ characteristicUUID: CBUUID) -> CBCharacteristic? {
@@ -229,7 +234,7 @@ class BLEPeripheral: RCTEventEmitter, CBPeripheralManagerDelegate {
     override func startObserving() { hasListeners = true }
     override func stopObserving() { hasListeners = false }
     @objc override static func requiresMainQueueSetup() -> Bool { return false }
-    
+
 }
 
 @available(iOS 10.0, *)
